+++ examples/shared/src/examples/Annotations/CustomCallout.tsx
@@ -3,7 +3,7 @@ import {
   MapView,
   MarkerView,
   GeoJSONSource,
-  SymbolLayer,
+  Layer,
 } from "@maplibre/maplibre-react-native";
 import { useState } from "react";
 import { Text, View } from "react-native";
@@ -28,7 +28,8 @@ export function CustomCallout() {
           setSelectedFeature(feature);
         }}
       >
-        <SymbolLayer
+        <Layer
+          type="symbol"
           id="symbol-layer"
           style={{
             iconAllowOverlap: true,