+++ ios/components/images/MLRNImages.h
@@ -10,15 +10,20 @@
 
 @property (nonatomic, weak) RCTBridge *bridge;
 
-@property (nonatomic, strong) MLRNMapView *map;
+@property (nonatomic, strong) MLRNMapView * _Nullable map;
+@property (nonatomic, strong, nonnull) NSMutableArray<id<RCTComponent>> *reactSubviews;
 
-@property (nonatomic, strong) NSDictionary<NSString *, NSString *> *images;
-@property (nonatomic, strong) NSArray<NSString *> *nativeImages;
+@property (nonatomic, strong, nonnull) NSDictionary * images;
 
-@property (nonatomic, copy) RCTBubblingEventBlock onImageMissing;
-@property (nonatomic, assign) BOOL hasOnImageMissing;
+@property (nonatomic, copy, nullable) RCTDirectEventBlock onImageMissing;
 
-- (BOOL)addMissingImageToStyle:(NSString *)imageName;
-- (void)sendImageMissingEvent:(NSString *)imageName;
+- (void)addToMap;
+- (void)removeFromMap;
+
+- (BOOL)addMissingImageToStyle:(NSString *_Nonnull)imageName;
+- (void)sendImageMissingEvent:(NSString *_Nonnull)imageName;
+
+- (void)insertReactSubview:(id<RCTComponent>_Nullable)subview atIndex:(NSInteger)atIndex;
+- (void)removeReactSubview:(id<RCTComponent>_Nullable)subview;
 
 @end