+++ ios/components/layers/MLRNBackgroundLayerManager.h
@@ -1,5 +1,0 @@
-#import "ViewManager.h"
-
-@interface MLRNBackgroundLayerManager : ViewManager
-
-@end