+++ src/index.ts
@@ -29,7 +29,7 @@ export {
   type PointAnnotationRef,
   type AnnotationEvent,
 } from "./components/annotations/PointAnnotation";
-export { Annotation } from "./components/annotations/Annotation";
+export { GeoJSONSourceAnnotation } from "./components/annotations/GeoJSONSourceAnnotation";
 export { Callout, type CalloutProps } from "./components/annotations/Callout";
 
 export { UserLocation } from "./components/user-location/UserLocation";
@@ -58,15 +58,28 @@ export {
   VectorSource,
 } from "./components/sources/vector-source/VectorSource";
 
-export { Images } from "./components/Images";
-export { FillLayer } from "./components/layers/FillLayer";
-export { FillExtrusionLayer } from "./components/layers/FillExtrusionLayer";
-export { HeatmapLayer } from "./components/layers/HeatmapLayer";
-export { LineLayer } from "./components/layers/LineLayer";
-export { CircleLayer } from "./components/layers/CircleLayer";
-export { SymbolLayer } from "./components/layers/SymbolLayer";
-export { RasterLayer } from "./components/layers/RasterLayer";
-export { BackgroundLayer } from "./components/layers/BackgroundLayer";
+export {
+  Layer,
+  type LayerProps,
+  type SourceLayerProps,
+  type StandaloneLayerProps,
+  type FillLayerProps,
+  type LineLayerProps,
+  type SymbolLayerProps,
+  type CircleLayerProps,
+  type HeatmapLayerProps,
+  type FillExtrusionLayerProps,
+  type RasterLayerProps,
+  type BackgroundLayerProps,
+} from "./components/layers/Layer";
+
+export {
+  type ImageSourceWithSdf,
+  type ImageEntry,
+  type ImagesProps,
+  Images,
+} from "./components/images/Images";
+
 export {
   MarkerView,
   type MarkerViewProps,